requirements.txt
appdirs==1.4.4
bcrypt==3.2.0
blinker==1.4
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
distlib==0.3.2
filelock==3.0.12
Flask==1.1.2
Flask-Mail==0.9.1
google-auth==1.32.1
google-auth-oauthlib==0.4.2
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
virtualenv==20.4.7
Werkzeug==1.0.1
